Removing BrowserProtocol as was not being used
[Tree Manager]
Added a pathObserving feature.
Observers must comply with PathObserverProtocol

[BrowserController]
added notification for the flatView changed event
Correction of the FoldersDisplayedMacro. Former implementation made it impossible to view Folders.
Doesn't ask for the hashObject to non TreeItem type classes

[TreeClasses]
Added a deinit selector, in all Tree Classes.
TreeBranchCatalyst addTreeItem always use TreeBranchCatalyst class. This is made consistent in all SubClasses (TreeCollection and filterBranch)
Added removeFilesWithoutDuplicate to TreeBranchCatalyst class
Added a constructor to the TreeBranchCatalyst class

[Duplicate Finder]
What determines the view, is actually if the tree is displayed or not. If the tree is displayed, it means that that a tree structure is being used. Otherwise, it displays a flat view, organised by duplicate number.
Added observes to pathChanges in both the flat and tree views
Refresh of the views when the files are deleted.