[All] Updating TODO: versions Removing Debug NSLogs Removing kSenderKey as it was not being used
AppDelegate:
Fixing problem in the moving method, it was calling a copy instead.
Start Duplicates now doesn't start busy animations on the main windows.

DuplicateFind:
Progress is reported in % of work done.
Adding protection to the setDuplicateCount in TreeLeaf.  What trying to set a null pointer
Increased the size of the table on the Duplicate settings

FileUtils:
renamed fileSize in exact_size. The former is being defined by the system elsewhere and was giving problems when used with (id) pointers.

AppStore Manager:
invalidating previous buys when refreshing receipts from the AppStore.
Replacing com.cascode.duplicates by com.cascode.duplicates2