[All] Updated to Xcode 8 Revised Icon View to support new technology. No Bindings to an array controller which made it heavy.
[BrowserController]
harmonizing the calls to the SetPathBar. Now it always goes through the SetCurrentNode.
Removing restrictions on the IconView, that was blocking the Flat Modes. (Let's see if it goes better this way)

[NodeViewController]
Now the dummy parent that represents the .. (parent folder) is always set as read-only.

[TableViewController]
the TextColor is coming directly from the TreeItem

[TreeItem]
Adding the textColor selector
Fixed pathLevel selector