[BrowserView] adding token Field
[BrowserController]
Correcting references to the old NSMutableArray that was used for managing roots for a treeCollection

[NodeViewController]
starting to change the grouping strategy.
All the access to the items is managed by these class. IconViewControllers and TableViewControllers access information through this class selectors

[TreeClasses]
Adding template like references on the arrays.
Adding a enumerator for the harvest of TreeItems (Not sure this is really efficient, but it is now working)
Reordering the class hierarchy